+++ package.json
@@ -27,7 +27,6 @@
     "@fleekhq/fleek-storage-js": "^1.0.20",
     "@hookform/devtools": "^4.0.1",
     "@makerdao/multicall": "^0.12.0",
-    "@metamask/jazzicon": "^2.0.0",
     "@metamask/onboarding": "^1.0.1",
     "@web3-react/core": "^6.1.9",
     "@web3-react/injected-connector": "^6.0.7",
