+++ src/components/[guild]/RequirementCard/RequirementCard.tsx
@@ -1,9 +1,9 @@
 import ColorCard from "components/common/ColorCard"
 import Link from "components/common/Link"
-import isNumber from "components/common/utils/isNumber"
 import RequirementChainTypeText from "components/create-guild/Requirements/components/RequirementChainTypeText"
 import { RPC } from "connectors"
 import { Requirement, RequirementTypeColors, Rest } from "types"
+import isNumber from "utils/isNumber"
 import MirrorEdition from "./components/MirrorEdition"
 import RequirementText from "./components/RequirementText"
 import SnapshotStrategy from "./components/SnapshotStrategy"
