+++ src/components/[guild]/RolesByPlatform/components/RoleListItem/RoleListItem.tsx
@@ -19,7 +19,7 @@ import useIsOwner from "components/[guild]/hooks/useIsOwner"
 import LogicDivider from "components/[guild]/LogicDivider"
 import RequirementCard from "components/[guild]/RequirementCard"
 import useLevelsAccess from "components/[guild]/RolesByPlatform/components/JoinButton/hooks/useLevelsAccess"
-import useRequirementLabels from "hooks/useRequirementLabels"
+import useRequirementLabels from "components/[guild]/RolesByPlatform/components/RoleListItem/hooks/useRequirementLabels"
 import dynamic from "next/dynamic"
 import { CaretDown, CaretUp, Check, X } from "phosphor-react"
 import React, { useState } from "react"
