+++ src/components/common/Layout/components/Footer/components/DotDelimiter.tsx
