--- src/components/common/Layout/components/InfoMenu/index.ts
@@ -1,3 +1,0 @@
-import InfoMenu from "./InfoMenu"
-
-export default InfoMenu
