--- src/constants/metadata/bayc.ts
@@ -1,186 +1,0 @@
-const bayc = {
-  background: [
-    "Purple",
-    "Yellow",
-    "Orange",
-    "Aquamarine",
-    "Army Green",
-    "Blue",
-    "New Punk Blue",
-    "Gray",
-  ],
-  clothes: [
-    "Striped Tee",
-    "Black T",
-    "Navy Striped Tee",
-    "Sailor Shirt",
-    "Hawaiian",
-    "Biker Vest",
-    "Sleeveless T",
-    "Wool Turtleneck",
-    "Prison Jumpsuit",
-    "Tanktop",
-    "Tuxedo Tee",
-    "Guayabera",
-    "Bone Tee",
-    "Puffy Vest",
-    "Vietnam Jacket",
-    "Smoking Jacket",
-    "Bayc T Black",
-    "Lumberjack Shirt",
-    "Leather Jacket",
-    "Black Holes T",
-    "Bone Necklace",
-    "Toga",
-    "Work Vest",
-    "Stunt Jacket",
-    "Bandolier",
-    "Caveman Pelt",
-    "Leather Punk Jacket",
-    "Lab Coat",
-    "Sleeveless Logo T",
-    "Tie Dye",
-    "Service",
-    "Tweed Suit",
-    "Bayc T Red",
-    "Rainbow Suspenders",
-    "Hip Hop",
-    "Cowboy Shirt",
-    "Space Suit",
-    "Prom Dress",
-    "Blue Dress",
-    "Pimp Coat",
-    "Kings Robe",
-    "Admirals Coat",
-    "Black Suit",
-  ],
-  earring: [
-    "Silver Hoop",
-    "Silver Stud",
-    "Gold Hoop",
-    "Gold Stud",
-    "Diamond Stud",
-    "Cross",
-  ],
-  eyes: [
-    "Bored",
-    "Bloodshot",
-    "Sleepy",
-    "Closed",
-    "Sad",
-    "Wide Eyed",
-    "3d",
-    "Coins",
-    "Angry",
-    "Crazy",
-    "Heart",
-    "Sunglasses",
-    "Robot",
-    "Eyepatch",
-    "Zombie",
-    "Blindfold",
-    "X Eyes",
-    "Scumbag",
-    "Hypnotized",
-    "Holographic",
-    "Cyborg",
-    "Laser Eyes",
-    "Blue Beams",
-  ],
-  fur: [
-    "Brown",
-    "Dark Brown",
-    "Black",
-    "Golden Brown",
-    "Cream",
-    "Tan",
-    "Pink",
-    "Gray",
-    "Blue",
-    "Red",
-    "Cheetah",
-    "White",
-    "Zombie",
-    "Robot",
-    "Dmt",
-    "Death Bot",
-    "Noise",
-    "Trippy",
-    "Solid Gold",
-  ],
-  hat: [
-    "Beanie",
-    "Seaman's Hat",
-    "Fez",
-    "Cowboy Hat",
-    "Fisherman's Hat",
-    "Halo",
-    "Short Mohawk",
-    "Commie Hat",
-    "Sea Captain's Hat",
-    "Army Hat",
-    "Bowler",
-    "Horns",
-    "S&m Hat",
-    "Bayc Flipped Brim",
-    "Bayc Hat Black",
-    "Irish Boho",
-    "Vietnam Era Helmet",
-    "Bunny Ears",
-    "Sushi Chef Headband",
-    "Safari",
-    "Spinner Hat",
-    "Baby's Bonnet",
-    "Stuntman Helmet",
-    "Girl's Hair Short",
-    "Faux Hawk",
-    "Police Motorcycle Helmet",
-    "Prussian Helmet",
-    "Party Hat 1",
-    "Bayc Hat Red",
-    "Ww2 Pilot Helm",
-    "Party Hat 2",
-    "Girl's Hair Pink",
-    "Bandana Blue",
-    "King's Crown",
-    "Laurel Wreath",
-    "Trippy Captain's Hat",
-  ],
-  mouth: [
-    "Bored",
-    "Bored Unshaven",
-    "Grin",
-    "Bored Cigarette",
-    "Dumbfounded",
-    "Bored Unshaven Cigarette",
-    "Phoneme Vuh",
-    "Jovial",
-    "Small Grin",
-    "Rage",
-    "Phoneme ooo",
-    "Phoneme L",
-    "Phoneme Oh",
-    "Discomfort",
-    "Tongue Out",
-    "Phoneme Wah",
-    "Bored Pipe",
-    "Bored Cigar",
-    "Bored Bubblegum",
-    "Grin Multicolored",
-    "Bored Unshaven Pipe",
-    "Bored Unshaven Cigar",
-    "Grin Gold Grill",
-    "Bored Party Horn",
-    "Grin Diamond Grill",
-    "Bored Kazoo",
-    "Bored Unshaven Bubblegum",
-    "Bored Unshaven Kazoo",
-    "Bored Pizza",
-    "Bored Dagger",
-    "Bored Unshaven Party horn",
-    "Bored Unshaven Dagger",
-    "Bored Unshaven Pizza",
-  ],
-}
-
-export default bayc
