--- src/constants/metadata/coolcats.ts
@@ -1,308 +1,0 @@
-const coolcats = {
-  body: ["Blue cat skin"],
-
-  face: [
-    "Overalls blue",
-    "Buttondown green",
-    "Robe blue",
-    "Bandana red",
-    "Bandana purple",
-    "Buttondown black flannel",
-    "Work red",
-    "Tanktop white",
-    "Tshirt pink",
-    "Bandana green",
-    "Shirt yellow",
-    "Tshirt yellow",
-    "Robe red",
-    "Tshirt red",
-    "Work blue",
-    "Tanktop pink",
-    "Buttondown blue flannel",
-    "Hoodie purple",
-    "Tanktop orange",
-    "Tshirt green",
-    "Sweater orange",
-    "Buttondown tan",
-    "Shirt white",
-    "Tshirt blue",
-    "Ninja blue",
-    "Tanktop tattoo",
-    "Overalls flannel",
-    "Ninja red",
-    "Labcoat",
-    "Baseball blue",
-    "Winter red",
-    "Sweater pink",
-    "Tanktop sailor red",
-    "Shirt bowtie",
-    "Tshirt metal",
-    "Tanktop sailor blue",
-    "Toga",
-    "Hoodie black",
-    "Tanktop sailor black",
-    "Buttondown red flannel",
-    "Robe white",
-    "Winter blue",
-    "Baseball red",
-    "Chain",
-    "Tshirt white",
-    "Sweater black",
-    "Overalls pink",
-    "Overalls yellow",
-    "Viking navy",
-    "Pirate black",
-    "Viking brown",
-    "Lederhosen",
-    "Astro",
-    "Hoodie red",
-    "Knight leather",
-    "Wetsuit",
-    "Cowboy black",
-    "Ninja black",
-    "Overalls red",
-    "Gown purple",
-    "Sweater green chain",
-    "Knight",
-    "Gown white",
-    "Pirate red",
-    "Epaulette black",
-    "Epaulette red",
-    "Costume hotdog",
-    "Punk",
-    "Combat black",
-    "Monk",
-    "Costume frog",
-    "Tiger",
-    "Astro black",
-    "Astro orange",
-    "Nurse",
-    "Deepsea bronze",
-    "Epaulette white",
-    "Robe king",
-    "Combat green",
-    "Cowboy brown",
-    "Gown black",
-    "Costume dragon",
-    "Mononoke",
-    "Deepsea orange",
-    "Knight black",
-    "Costume gorilla",
-    "Alien",
-    "Angel",
-    "Celestial",
-    "Demon",
-    "Lucky",
-    "Robot",
-    "Skeleton",
-    "Special zombie",
-    "Upsidedown",
-  ],
-
-  hats: [
-    "Hat white",
-    "Bucket hat green",
-    "Mohawk green",
-    "Antlers",
-    "Bucket hat tan",
-    "Apple",
-    "Headband blue",
-    "Hat visor yellow",
-    "Beret green",
-    "Flower red",
-    "Candle",
-    "Mohawk purple",
-    "Headband red",
-    "Beanie red",
-    "Piercings",
-    "Flower pink",
-    "Arrowhead",
-    "Beret red",
-    "Flower blue",
-    "Wreath",
-    "Beanie orange",
-    "Cowboy brown",
-    "Sunhat tan",
-    "Hat skull",
-    "Mohawk red",
-    "Visor green",
-    "Beret black",
-    "Beret pink",
-    "Cupcake",
-    "Afro brown",
-    "Pirate red",
-    "Hat visor blue",
-    "Ninja blue",
-    "Sunhat black",
-    "Beanie black",
-    "Bucket hat white",
-    "Helm army",
-    "Sunhat white",
-    "Unicorn horn",
-    "Afro black",
-    "Mullet brown",
-    "Mullet blonde",
-    "Beanie blue",
-    "Hat black",
-    "Sushi",
-    "Cowboy black",
-    "Knight red",
-    "Ninja black",
-    "Ninja red",
-    "Wreath flowers",
-    "Bow",
-    "Prince",
-    "Tvhead grey",
-    "Nurse",
-    "Pirate black",
-    "Halo fire",
-    "Horns",
-    "Top hat",
-    "Bucket hat blue",
-    "Halo",
-    "Helm biker",
-    "Goggles seaweed",
-    "Knight blue",
-    "Astro",
-    "Admiral",
-    "Tvhead white",
-    "Tvhead purple",
-    "Astro cheeks",
-    "Helm bronze",
-    "Visor purple",
-    "Costume gorilla",
-    "Dutch",
-    "Helm silver",
-    "Admiral pink",
-    "Crown black",
-    "Crown gold",
-    "Afro rainbow unicorn",
-    "Crown fire",
-    "Deepsea bronze",
-    "Costume dragon",
-    "Costume frog",
-    "Deepsea orange",
-    "Knight black",
-    "Astro fishbowl",
-    "Alien",
-    "Angel",
-    "Celestial",
-    "Demon",
-    "Lucky",
-    "Robot",
-    "Skeleton",
-    "Special zombie",
-    "Upsidedown",
-  ],
-
-  shirt: [
-    "Overalls blue",
-    "Buttondown green",
-    "Robe blue",
-    "Bandana red",
-    "Bandana purple",
-    "Buttondown black flannel",
-    "Work red",
-    "Tanktop white",
-    "Tshirt pink",
-    "Bandana green",
-    "Shirt yellow",
-    "Tshirt yellow",
-    "Robe red",
-    "Tshirt red",
-    "Work blue",
-    "Tanktop pink",
-    "Buttondown blue flannel",
-    "Hoodie purple",
-    "Tanktop orange",
-    "Tshirt green",
-    "Sweater orange",
-    "Buttondown tan",
-    "Shirt white",
-    "Tshirt blue",
-    "Ninja blue",
-    "Tanktop tattoo",
-    "Overalls flannel",
-    "Ninja red",
-    "Labcoat",
-    "Baseball blue",
-    "Winter red",
-    "Sweater pink",
-    "Tanktop sailor red",
-    "Shirt bowtie",
-    "Tshirt metal",
-    "Tanktop sailor blue",
-    "Toga",
-    "Hoodie black",
-    "Tanktop sailor black",
-    "Buttondown red flannel",
-    "Robe white",
-    "Winter blue",
-    "Baseball red",
-    "Chain",
-    "Tshirt white",
-    "Sweater black",
-    "Overalls pink",
-    "Overalls yellow",
-    "Viking navy",
-    "Pirate black",
-    "Viking brown",
-    "Lederhosen",
-    "Astro",
-    "Hoodie red",
-    "Knight leather",
-    "Wetsuit",
-    "Cowboy black",
-    "Ninja black",
-    "Overalls red",
-    "Gown purple",
-    "Sweater green chain",
-    "Knight",
-    "Gown white",
-    "Pirate red",
-    "Epaulette black",
-    "Epaulette red",
-    "Costume hotdog",
-    "Punk",
-    "Combat black",
-    "Monk",
-    "Costume frog",
-    "Tiger",
-    "Astro black",
-    "Astro orange",
-    "Nurse",
-    "Deepsea bronze",
-    "Epaulette white",
-    "Robe king",
-    "Combat green",
-    "Cowboy brown",
-    "Gown black",
-    "Costume dragon",
-    "Mononoke",
-    "Deepsea orange",
-    "Knight black",
-    "Costume gorilla",
-    "Alien",
-    "Angel",
-    "Celestial",
-    "Demon",
-    "Lucky",
-    "Robot",
-    "Skeleton",
-    "Special zombie",
-    "Upsidedown",
-  ],
-
-  tier: [
-    "Cool 2",
-    "Cool 1",
-    "Wild 1",
-    "Wild 2",
-    "Classy 1",
-    "Classy 2",
-    "Exotic 1",
-    "Exotic 2",
-  ],
-}
-
-export default coolcats
