--- src/constants/metadata/cryptopunks.ts
@@ -1,94 +1,0 @@
-const cryptopunks = {
-  accessory: [
-    "Earring",
-    "Cigarette",
-    "Hot Lipstick",
-    "Purple Lipstick",
-    "Mole",
-    "Black Lipstick",
-    "Nerd Glasses",
-    "Big Shades",
-    "Horned Rim Glasses",
-    "Regular Shades",
-    "Shadow Beard",
-    "Classic Shades",
-    "Bandana",
-    "Stringy Hair",
-    "Eye Patch",
-    "Messy Hair",
-    "Wild Hair",
-    "Frumpy Hair",
-    "Mohawk",
-    "Mohawk Thin",
-    "Mohawk Dark",
-    "Knitted Cap",
-    "Crazy Hair",
-    "Headband",
-    "Clown Eyes Blue",
-    "Clown Eyes Green",
-    "Small Shades",
-    "Cap",
-    "VR",
-    "Pipe",
-    "Muttonchops",
-    "Peak Spike",
-    "Do-rag",
-    "Shaved Head",
-    "Goat",
-    "Eye Mask",
-    "Normal Beard",
-    "Normal Beard Black",
-    "Mustache",
-    "3D Glasses",
-    "Luxurious Beard",
-    "Chinstrap",
-    "Front Beard",
-    "Vape",
-    "Green Eye Shadow",
-    "Blue Eye Shadow",
-    "Handlebars",
-    "Purple Eye Shadow",
-    "Frown",
-    "Front Beard Dark",
-    "Hoodie",
-    "Cap Forward",
-    "Smile",
-    "Clown Nose",
-    "Police Cap",
-    "Fedora",
-    "Tassle Hat",
-    "Medical Mask",
-    "Gold Chain",
-    "Purple Hair",
-    "Dark Hair",
-    "Silver Chain",
-    "Straight Hair",
-    "Clown Hair Green",
-    "Straight Hair Dark",
-    "Blonde Bob",
-    "Half Shaved",
-    "Red Mohawk",
-    "Vampire Hair",
-    "Big Beard",
-    "Straight Hair Blonde",
-    "Wild Blonde",
-    "Cowboy Hat",
-    "Wild White Hair",
-    "Blonde Short",
-    "Rosy Cheeks",
-    "Spots",
-    "Top Hat",
-    "Pink With Hat",
-    "Pigtails",
-    "Welding Goggles",
-    "Buck Teeth",
-    "Orange Side",
-    "Tiara",
-    "Pilot Helmet",
-    "Choker",
-    "Beanie",
-  ],
-  type: ["Male", "Female", "Zombie", "Ape", "Alien"],
-}
-
-export default cryptopunks
