--- src/constants/metadata/index.ts
@@ -1,9 +1,0 @@
-import BAYC from "./bayc"
-import COOLCATS from "./coolcats"
-import CRYPTOPUNKS from "./cryptopunks"
-import LOOT from "./loot"
-import MUTAGEN from "./mutagen"
-
-const metadata = { BAYC, CRYPTOPUNKS, LOOT, COOLCATS, MUTAGEN }
-
-export default metadata
