--- src/constants/metadata/loot.ts
@@ -1,827 +1,0 @@
-const loot = {
-  chest: [
-    `Linen Robe`,
-    `Studded Leather Armor`,
-    `Divine Robe`,
-    `Robe`,
-    `Leather Armor`,
-    `Chain Mail`,
-    `Ornate Chestplate`,
-    `Plate Mail`,
-    `Hard Leather Armor`,
-    `Shirt`,
-    `Demon Husk`,
-    `Holy Chestplate`,
-    `Ring Mail`,
-    `Silk Robe`,
-    `Dragonskin Armor`,
-    `Plate Mail of Brilliance`,
-    `Robe of Anger`,
-    `Leather Armor of Protection`,
-    `Plate Mail of Reflection`,
-    `Divine Robe of Rage`,
-    `Leather Armor of Titans`,
-    `Leather Armor of the Fox`,
-    `Plate Mail of Detection`,
-    `Plate Mail of Giants`,
-    `Dragonskin Armor of Fury`,
-    `Dragonskin Armor of Power`,
-    `Leather Armor of Perfection`,
-    `Leather Armor of Skill`,
-    `Plate Mail of Fury`,
-    `Plate Mail of Rage`,
-    `Dragonskin Armor of Enlightenment`,
-    `Dragonskin Armor of Giants`,
-    `Dragonskin Armor of Perfection`,
-    `Dragonskin Armor of Protection`,
-    `Plate Mail of Power`,
-    `Plate Mail of Vitriol`,
-    `Ring Mail of Giants`,
-    `Robe of Titans`,
-    `Divine Robe of Protection`,
-    `Dragonskin Armor of the Fox`,
-    `Leather Armor of Vitriol`,
-    `Ornate Chestplate of Detection`,
-    `Robe of Brilliance`,
-    `Robe of Enlightenment`,
-    `Robe of Power`,
-    `Chain Mail of Protection`,
-    `Divine Robe of Fury`,
-    `Divine Robe of Giants`,
-    `Divine Robe of Titans`,
-    `Divine Robe of Vitriol`,
-    `Dragonskin Armor of Rage`,
-    `Leather Armor of Anger`,
-    `Leather Armor of Rage`,
-    `Plate Mail of Protection`,
-    `Robe of Vitriol`,
-    `Studded Leather Armor of Giants`,
-    `Divine Robe of Brilliance`,
-    `Divine Robe of Perfection`,
-    `Hard Leather Armor of Detection`,
-    `Holy Chestplate of Titans`,
-    `Leather Armor of Giants`,
-    `Leather Armor of the Twins`,
-    `Plate Mail of the Fox`,
-    `Plate Mail of the Twins`,
-    `Robe of Reflection`,
-    `Robe of Skill`,
-    `Robe of the Twins`,
-    `Chain Mail of Giants`,
-    `Divine Robe of Enlightenment`,
-    `Divine Robe of Reflection`,
-    `Dragonskin Armor of Detection`,
-    `Holy Chestplate of Brilliance`,
-    `Leather Armor of Power`,
-    `Linen Robe of Rage`,
-    `Robe of Detection`,
-    `Robe of Fury`,
-    `Silk Robe of the Twins`,
-    `Chain Mail of Anger`,
-    `Chain Mail of Rage`,
-    `Divine Robe of the Twins`,
-    `Dragonskin Armor of Anger`,
-    `Dragonskin Armor of Brilliance`,
-    `Hard Leather Armor of Enlightenment`,
-    `Holy Chestplate of the Fox`,
-    `Linen Robe of Vitriol`,
-    `Plate Mail of Anger`,
-    `Ring Mail of Anger`,
-    `Ring Mail of the Fox`,
-    `Robe of Giants`,
-    `Robe of Protection`,
-    `Robe of Rage`,
-    `Shirt of Rage`,
-    `Silk Robe of Perfection`,
-    `Studded Leather Armor of Detection`,
-    `Studded Leather Armor of the Twins`,
-    `Demon Husk of Protection`,
-    `Hard Leather Armor of Anger`,
-    `Leather Armor of Enlightenment`,
-    `Linen Robe of Protection`,
-    `Ring Mail of Fury`,
-  ],
-
-  foot: [
-    `Chain Boots`,
-    `Linen Shoes`,
-    `Divine Slippers`,
-    `Greaves`,
-    `Wool Shoes`,
-    `Shoes`,
-    `Studded Leather Boots`,
-    `Leather Boots`,
-    `Ornate Greaves`,
-    `Demonhide Boots`,
-    `Holy Greaves`,
-    `Dragonskin Boots`,
-    `Silk Slippers`,
-    `Hard Leather Boots`,
-    `Heavy Boots`,
-    `Dragonskin Boots of Detection`,
-    `Dragonskin Boots of Fury`,
-    `Chain Boots of Detection`,
-    `Leather Boots of Detection`,
-    `Leather Boots of Enlightenment`,
-    `Wool Shoes of Giants`,
-    `Chain Boots of Perfection`,
-    `Dragonskin Boots of Rage`,
-    `Dragonskin Boots of Anger`,
-    `Holy Greaves of Fury`,
-    `Holy Greaves of Perfection`,
-    `Holy Greaves of Skill`,
-    `Leather Boots of Brilliance`,
-    `Leather Boots of Protection`,
-    `Wool Shoes of Vitriol`,
-    `Wool Shoes of the Twins`,
-    `Chain Boots of Reflection`,
-    `Dragonskin Boots of Reflection`,
-    `Holy Greaves of Enlightenment`,
-    `Holy Greaves of Power`,
-    `Leather Boots of the Fox`,
-    `Shoes of Brilliance`,
-    `Chain Boots of Anger`,
-    `Chain Boots of Fury`,
-    `Holy Greaves of Vitriol`,
-    `Leather Boots of Anger`,
-    `Leather Boots of Fury`,
-    `Wool Shoes of Anger`,
-    `Wool Shoes of Rage`,
-    `Wool Shoes of the Fox`,
-    `Chain Boots of Enlightenment`,
-    `Chain Boots of Giants`,
-    `Chain Boots of Protection`,
-    `Chain Boots of Skill`,
-    `Chain Boots of the Fox`,
-    `Dragonskin Boots of Enlightenment`,
-    `Dragonskin Boots of Titans`,
-    `Heavy Boots of Brilliance`,
-    `Wool Shoes of Brilliance`,
-    `Chain Boots of Rage`,
-    `Chain Boots of Titans`,
-    `Chain Boots of Vitriol`,
-    `Dragonskin Boots of Brilliance`,
-    `Dragonskin Boots of Giants`,
-    `Dragonskin Boots of Perfection`,
-    `Dragonskin Boots of Protection`,
-    `Dragonskin Boots of Vitriol`,
-    `Dragonskin Boots of the Twins`,
-    `Holy Greaves of Anger`,
-    `Holy Greaves of Brilliance`,
-    `Holy Greaves of Rage`,
-    `Holy Greaves of Reflection`,
-    `Holy Greaves of Titans`,
-    `Leather Boots of Vitriol`,
-    `Ornate Greaves of the Fox`,
-    `Wool Shoes of Enlightenment`,
-    `Wool Shoes of Fury`,
-    `Wool Shoes of Perfection`,
-    `Chain Boots of Brilliance`,
-    `Chain Boots of Power`,
-    `Demonhide Boots of Detection`,
-    `Greaves of Perfection`,
-    `Holy Greaves of Protection`,
-    `Holy Greaves of the Fox`,
-    `Holy Greaves of the Twins`,
-    `Leather Boots of Giants`,
-    `Leather Boots of Perfection`,
-    `Leather Boots of the Twins`,
-    `Ornate Greaves of Fury`,
-    `Silk Slippers of Detection`,
-    `Studded Leather Boots of Brilliance`,
-    `Wool Shoes of Power`,
-    `Wool Shoes of Protection`,
-    `Wool Shoes of Reflection`,
-    `Demonhide Boots of Brilliance`,
-    `Divine Slippers of Anger`,
-    `Divine Slippers of Giants`,
-    `Divine Slippers of the Twins`,
-    `Heavy Boots of Protection`,
-    `Leather Boots of Skill`,
-    `Linen Shoes of Rage`,
-    `Ornate Greaves of Titans`,
-    `Silk Slippers of Enlightenment`,
-    `Wool Shoes of Skill`,
-    `Wool Shoes of Titans`,
-  ],
-
-  hand: [
-    `Studded Leather Gloves`,
-    `Silk Gloves`,
-    `Heavy Gloves`,
-    `Linen Gloves`,
-    `Chain Gloves`,
-    `Dragonskin Gloves`,
-    `Wool Gloves`,
-    `Holy Gauntlets`,
-    `Gauntlets`,
-    `Divine Gloves`,
-    `Leather Gloves`,
-    `Gloves`,
-    `Ornate Gauntlets`,
-    `Hard Leather Gloves`,
-    `Demon's Hands`,
-    `Holy Gauntlets of Giants`,
-    `Chain Gloves of Perfection`,
-    `Leather Gloves of Anger`,
-    `Wool Gloves of Titans`,
-    `Holy Gauntlets of Perfection`,
-    `Wool Gloves of the Twins`,
-    `Dragonskin Gloves of the Twins`,
-    `Holy Gauntlets of Detection`,
-    `Holy Gauntlets of Skill`,
-    `Holy Gauntlets of Titans`,
-    `Leather Gloves of the Fox`,
-    `Wool Gloves of Anger`,
-    `Wool Gloves of Detection`,
-    `Holy Gauntlets of Rage`,
-    `Holy Gauntlets of Vitriol`,
-    `Leather Gloves of Enlightenment`,
-    `Leather Gloves of Giants`,
-    `Leather Gloves of Protection`,
-    `Leather Gloves of Titans`,
-    `Wool Gloves of Giants`,
-    `Wool Gloves of Protection`,
-    `Chain Gloves of Anger`,
-    `Dragonskin Gloves of Power`,
-    `Dragonskin Gloves of Rage`,
-    `Dragonskin Gloves of Vitriol`,
-    `Holy Gauntlets of the Twins`,
-    `Leather Gloves of Perfection`,
-    `Leather Gloves of the Twins`,
-    `Wool Gloves of Skill`,
-    `Chain Gloves of Enlightenment`,
-    `Chain Gloves of Power`,
-    `Chain Gloves of Vitriol`,
-    `Dragonskin Gloves of Brilliance`,
-    `Dragonskin Gloves of Detection`,
-    `Dragonskin Gloves of Titans`,
-    `Dragonskin Gloves of the Fox`,
-    `Gauntlets of Anger`,
-    `Gloves of Brilliance`,
-    `Holy Gauntlets of Anger`,
-    `Holy Gauntlets of Fury`,
-    `Holy Gauntlets of Protection`,
-    `Holy Gauntlets of Reflection`,
-    `Leather Gloves of Brilliance`,
-    `Wool Gloves of Brilliance`,
-    `Wool Gloves of Reflection`,
-    `Chain Gloves of Giants`,
-    `Chain Gloves of the Fox`,
-    `Demon's Hands of Titans`,
-    `Dragonskin Gloves of Skill`,
-    `Gauntlets of Protection`,
-    `Hard Leather Gloves of Giants`,
-    `Hard Leather Gloves of Skill`,
-    `Hard Leather Gloves of the Twins`,
-    `Holy Gauntlets of Brilliance`,
-    `Holy Gauntlets of Power`,
-    `Leather Gloves of Rage`,
-    `Leather Gloves of Reflection`,
-    `Leather Gloves of Skill`,
-    `Linen Gloves of the Fox`,
-    `Silk Gloves of Vitriol`,
-    `Studded Leather Gloves of Fury`,
-    `Wool Gloves of Fury`,
-    `Wool Gloves of Perfection`,
-    `Wool Gloves of Power`,
-    `Chain Gloves of Rage`,
-    `Demon's Hands of Rage`,
-    `Divine Gloves of Fury`,
-    `Divine Gloves of Protection`,
-    `Dragonskin Gloves of Perfection`,
-    `Dragonskin Gloves of Reflection`,
-    `Gauntlets of Giants`,
-    `Hard Leather Gloves of Enlightenment`,
-    `Hard Leather Gloves of Rage`,
-    `Hard Leather Gloves of Reflection`,
-    `Heavy Gloves of Power`,
-    `Heavy Gloves of Skill`,
-    `Holy Gauntlets of the Fox`,
-    `Leather Gloves of Power`,
-    `Leather Gloves of Vitriol`,
-    `Linen Gloves of Fury`,
-    `Linen Gloves of Perfection`,
-    `Ornate Gauntlets of Perfection`,
-    `Ornate Gauntlets of the Fox`,
-    `Silk Gloves of Brilliance`,
-    `Silk Gloves of Giants`,
-  ],
-
-  head: [
-    `Hood`,
-    `Ornate Helm`,
-    `Great Helm`,
-    `Helm`,
-    `Linen Hood`,
-    `Divine Hood`,
-    `Crown`,
-    `Cap`,
-    `Ancient Helm`,
-    `Demon Crown`,
-    `Leather Cap`,
-    `Full Helm`,
-    `War Cap`,
-    `Dragon's Crown`,
-    `Silk Hood`,
-    `Cap of the Fox`,
-    `Full Helm of Detection`,
-    `Dragon's Crown of Fury`,
-    `Dragon's Crown of Detection`,
-    `Dragon's Crown of Protection`,
-    `Full Helm of Giants`,
-    `Ancient Helm of Reflection`,
-    `Full Helm of Vitriol`,
-    `Silk Hood of Brilliance`,
-    `Ancient Helm of Detection`,
-    `Ancient Helm of Fury`,
-    `Dragon's Crown of Reflection`,
-    `Silk Hood of Power`,
-    `Silk Hood of Reflection`,
-    `Ancient Helm of Protection`,
-    `Cap of Detection`,
-    `Cap of Giants`,
-    `Cap of Power`,
-    `Cap of Titans`,
-    `Cap of Vitriol`,
-    `Cap of the Twins`,
-    `Dragon's Crown of Enlightenment`,
-    `Dragon's Crown of Perfection`,
-    `Dragon's Crown of Power`,
-    `Dragon's Crown of the Twins`,
-    `Full Helm of Reflection`,
-    `Silk Hood of the Fox`,
-    `Ancient Helm of Anger`,
-    `Ancient Helm of Rage`,
-    `Cap of Perfection`,
-    `Dragon's Crown of Anger`,
-    `Dragon's Crown of Brilliance`,
-    `Dragon's Crown of Giants`,
-    `Dragon's Crown of Vitriol`,
-    `Full Helm of Brilliance`,
-    `Full Helm of Fury`,
-    `Great Helm of Anger`,
-    `Linen Hood of Anger`,
-    `Linen Hood of Perfection`,
-    `Silk Hood of Anger`,
-    `Silk Hood of Enlightenment`,
-    `Silk Hood of Skill`,
-    `Silk Hood of Titans`,
-    `Ancient Helm of Enlightenment`,
-    `Ancient Helm of Perfection`,
-    `Ancient Helm of the Fox`,
-    `Cap of Protection`,
-    `Cap of Rage`,
-    `Demon Crown of Rage`,
-    `Divine Hood of Protection`,
-    `Dragon's Crown of Rage`,
-    `Great Helm of Skill`,
-    `Leather Cap of Fury`,
-    `Silk Hood of Detection`,
-    `Silk Hood of Giants`,
-    `Silk Hood of the Twins`,
-    `Ancient Helm of Giants`,
-    `Ancient Helm of the Twins`,
-    `Cap of Fury`,
-    `Crown of Anger`,
-    `Crown of Perfection`,
-    `Demon Crown of Protection`,
-    `Divine Hood of Reflection`,
-    `Divine Hood of Titans`,
-    `Full Helm of Enlightenment`,
-    `Full Helm of Rage`,
-    `Great Helm of Perfection`,
-    `Leather Cap of Detection`,
-    `Leather Cap of the Twins`,
-    `Linen Hood of Brilliance`,
-    `Linen Hood of Giants`,
-    `Silk Hood of Fury`,
-    `War Cap of Detection`,
-    `Ancient Helm of Power`,
-    `Cap of Anger`,
-    `Cap of Brilliance`,
-    `Cap of Enlightenment`,
-    `Dragon's Crown of Skill`,
-    `Full Helm of Perfection`,
-    `Full Helm of Power`,
-    `Full Helm of Protection`,
-    `Full Helm of Skill`,
-    `Full Helm of Titans`,
-    `Great Helm of Giants`,
-    `Hood of Giants`,
-  ],
-
-  neck: [
-    `Pendant`,
-    `Necklace`,
-    `Amulet`,
-    `Necklace of Fury`,
-    `Necklace of Protection`,
-    `Necklace of Brilliance`,
-    `Necklace of Power`,
-    `Necklace of Skill`,
-    `Necklace of the Fox`,
-    `Necklace of Giants`,
-    `Necklace of Titans`,
-    `Necklace of Detection`,
-    `Necklace of Perfection`,
-    `Necklace of Enlightenment`,
-    `Necklace of Vitriol`,
-    `Necklace of Rage`,
-    `Necklace of Reflection`,
-    `Necklace of the Twins`,
-    `Necklace of Anger`,
-    `Pendant of Detection`,
-    `Pendant of Titans`,
-    `Amulet of Detection`,
-    `Amulet of Enlightenment`,
-    `Amulet of Titans`,
-    `Pendant of the Fox`,
-    `Amulet of Reflection`,
-    `Amulet of Skill`,
-    `Pendant of Giants`,
-    `Pendant of Enlightenment`,
-    `Pendant of Fury`,
-    `Amulet of Fury`,
-    `Pendant of Protection`,
-    `Pendant of Reflection`,
-    `Pendant of Skill`,
-    `Amulet of Giants`,
-    `Amulet of Vitriol`,
-    `Pendant of Anger`,
-    `Pendant of Power`,
-    `Pendant of Vitriol`,
-    `Amulet of Power`,
-    `Amulet of Rage`,
-    `Amulet of the Twins`,
-    `Pendant of Rage`,
-    `Pendant of Perfection`,
-    `Amulet of Anger`,
-    `Amulet of Perfection`,
-    `Amulet of Protection`,
-    `Amulet of the Fox`,
-    `Pendant of the Twins`,
-    `Pendant of Brilliance`,
-    `Amulet of Brilliance`,
-    `"Armageddon Peak" Pendant of Fury +1`,
-    `"Brimstone Moon" Pendant of Detection +1`,
-    `"Dread Bite" Pendant of Power +1`,
-    `"Dread Moon" Pendant of Rage +1`,
-    `"Grim Bite" Pendant of Perfection +1`,
-    `"Havoc Sun" Amulet of Reflection`,
-    `"Oblivion Bender" Pendant of Reflection +1`,
-    `"Pandemonium Bite" Pendant of Anger +1`,
-    `"Apocalypse Glow" Amulet of Brilliance`,
-    `"Apocalypse Sun" Amulet of Titans`,
-    `"Corruption Bender" Pendant of Titans +1`,
-    `"Death Glow" Amulet of Brilliance`,
-    `"Demon Shout" Pendant of Vitriol +1`,
-    `"Dragon Roar" Amulet of Anger`,
-    `"Dusk Roar" Amulet of Titans`,
-    `"Eagle Bite" Pendant of Reflection +1`,
-    `"Fate Sun" Amulet of Perfection`,
-    `"Fate Whisper" Amulet of Power`,
-    `"Ghoul Roar" Amulet of Fury`,
-    `"Ghoul Sun" Amulet of Reflection`,
-    `"Gloom Bender" Pendant of Titans +1`,
-    `"Golem Root" Amulet of Detection`,
-    `"Havoc Roar" Amulet of Power`,
-    `"Honour Moon" Pendant of Detection +1`,
-    `"Hypnotic Tear" Amulet of Vitriol`,
-    `"Kraken Bender" Pendant of Anger +1`,
-    `"Kraken Bite" Pendant of Reflection +1`,
-    `"Kraken Moon" Pendant of Protection +1`,
-    `"Maelstrom Roar" Amulet of Perfection`,
-    `"Morbid Glow" Amulet of the Twins`,
-    `"Morbid Root" Amulet of Vitriol`,
-    `"Morbid Sun" Amulet of Perfection`,
-    `"Morbid Tear" Amulet of Brilliance`,
-    `"Oblivion Peak" Pendant of the Fox +1`,
-    `"Plague Tear" Amulet of Rage`,
-    `"Rune Sun" Amulet of Enlightenment`,
-    `"Shimmering Bite" Pendant of Perfection +1`,
-    `"Shimmering Moon" Pendant of Detection +1`,
-    `"Shimmering Shout" Pendant of Vitriol +1`,
-    `"Skull Bite" Pendant of Perfection +1`,
-    `"Sorrow Moon" Pendant of Protection +1`,
-    `"Soul Roar" Amulet of Perfection`,
-    `"Tempest Moon" Pendant of Rage +1`,
-    `"Vengeance Roar" Amulet of Anger`,
-    `"Vengeance Root" Amulet of Detection`,
-    `"Vengeance Root" Amulet of Rage`,
-    `"Vortex Sun" Amulet of Reflection`,
-    `"Vortex Whisper" Amulet of Perfection`,
-    `"Woe Moon" Pendant of Brilliance +1`,
-  ],
-
-  ring: [
-    `Silver Ring`,
-    `Bronze Ring`,
-    `Platinum Ring`,
-    `Titanium Ring`,
-    `Gold Ring`,
-    `Titanium Ring of Power`,
-    `Silver Ring of Protection`,
-    `Platinum Ring of Anger`,
-    `Platinum Ring of Reflection`,
-    `Silver Ring of the Fox`,
-    `Titanium Ring of Brilliance`,
-    `Titanium Ring of Giants`,
-    `Titanium Ring of Titans`,
-    `Gold Ring of Giants`,
-    `Titanium Ring of Vitriol`,
-    `Bronze Ring of Vitriol`,
-    `Platinum Ring of Vitriol`,
-    `Silver Ring of Anger`,
-    `Bronze Ring of Protection`,
-    `Gold Ring of Power`,
-    `Silver Ring of Rage`,
-    `Silver Ring of the Twins`,
-    `Bronze Ring of Power`,
-    `Gold Ring of Fury`,
-    `Gold Ring of Perfection`,
-    `Platinum Ring of Brilliance`,
-    `Titanium Ring of Anger`,
-    `Bronze Ring of Perfection`,
-    `Gold Ring of Anger`,
-    `Platinum Ring of Detection`,
-    `Platinum Ring of Titans`,
-    `Platinum Ring of the Twins`,
-    `Silver Ring of Power`,
-    `Silver Ring of Reflection`,
-    `Titanium Ring of Fury`,
-    `Bronze Ring of the Fox`,
-    `Gold Ring of Vitriol`,
-    `Platinum Ring of Power`,
-    `Platinum Ring of Protection`,
-    `Bronze Ring of Fury`,
-    `Bronze Ring of Skill`,
-    `Platinum Ring of Enlightenment`,
-    `Platinum Ring of Giants`,
-    `Platinum Ring of the Fox`,
-    `Silver Ring of Fury`,
-    `Silver Ring of Titans`,
-    `Titanium Ring of Perfection`,
-    `Titanium Ring of Reflection`,
-    `Titanium Ring of Skill`,
-    `Titanium Ring of the Twins`,
-    `Bronze Ring of Detection`,
-    `Bronze Ring of Rage`,
-    `Gold Ring of Rage`,
-    `Gold Ring of Titans`,
-    `Gold Ring of the Fox`,
-    `Gold Ring of the Twins`,
-    `Platinum Ring of Skill`,
-    `Silver Ring of Giants`,
-    `Bronze Ring of Anger`,
-    `Bronze Ring of Brilliance`,
-    `Bronze Ring of Reflection`,
-    `Gold Ring of Skill`,
-    `Silver Ring of Skill`,
-    `Titanium Ring of Enlightenment`,
-    `Titanium Ring of Rage`,
-    `Titanium Ring of the Fox`,
-    `Bronze Ring of Enlightenment`,
-    `Bronze Ring of Giants`,
-    `Bronze Ring of the Twins`,
-    `Gold Ring of Brilliance`,
-    `Gold Ring of Reflection`,
-    `Platinum Ring of Rage`,
-    `Silver Ring of Detection`,
-    `Titanium Ring of Detection`,
-    `Gold Ring of Protection`,
-    `Silver Ring of Vitriol`,
-    `Bronze Ring of Titans`,
-    `Gold Ring of Detection`,
-    `Platinum Ring of Fury`,
-    `Platinum Ring of Perfection`,
-    `Silver Ring of Brilliance`,
-    `Silver Ring of Perfection`,
-    `Gold Ring of Enlightenment`,
-    `Titanium Ring of Protection`,
-    `Silver Ring of Enlightenment`,
-    `"Blight Bender" Gold Ring of Perfection +1`,
-    `"Blight Grasp" Titanium Ring of Giants +1`,
-    `"Brimstone Bite" Platinum Ring of Power +1`,
-    `"Corruption Peak" Silver Ring of the Fox +1`,
-    `"Dragon Glow" Platinum Ring of the Twins`,
-    `"Foe Peak" Silver Ring of Power +1`,
-    `"Maelstrom Root" Bronze Ring of Skill`,
-    `"Mind Bender" Silver Ring of Perfection +1`,
-    `"Morbid Tear" Platinum Ring of Rage`,
-    `"Oblivion Peak" Gold Ring of Perfection +1`,
-    `"Pain Glow" Platinum Ring of Detection`,
-    `"Plague Roar" Silver Ring of Enlightenment`,
-    `"Skull Moon" Bronze Ring of Detection +1`,
-    `"Sorrow Peak" Titanium Ring of Enlightenment +1`,
-    `"Woe Bender" Silver Ring of Anger +1`,
-  ],
-
-  waist: [
-    `War Belt`,
-    `Heavy Belt`,
-    `Wool Sash`,
-    `Silk Sash`,
-    `Linen Sash`,
-    `Dragonskin Belt`,
-    `Ornate Belt`,
-    `Plated Belt`,
-    `Brightsilk Sash`,
-    `Demonhide Belt`,
-    `Studded Leather Belt`,
-    `Leather Belt`,
-    `Sash`,
-    `Mesh Belt`,
-    `Hard Leather Belt`,
-    `Leather Belt of Reflection`,
-    `Wool Sash of Giants`,
-    `Dragonskin Belt of Enlightenment`,
-    `Ornate Belt of Perfection`,
-    `Leather Belt of Enlightenment`,
-    `Leather Belt of Power`,
-    `Wool Sash of Rage`,
-    `Leather Belt of Fury`,
-    `Leather Belt of Perfection`,
-    `Leather Belt of Vitriol`,
-    `Mesh Belt of Titans`,
-    `Ornate Belt of Fury`,
-    `Ornate Belt of Skill`,
-    `Wool Sash of Skill`,
-    `Dragonskin Belt of Anger`,
-    `Dragonskin Belt of Protection`,
-    `Dragonskin Belt of Vitriol`,
-    `Leather Belt of Anger`,
-    `Leather Belt of Giants`,
-    `Mesh Belt of Brilliance`,
-    `Ornate Belt of Brilliance`,
-    `Ornate Belt of Detection`,
-    `Wool Sash of Protection`,
-    `Wool Sash of Titans`,
-    `Wool Sash of the Twins`,
-    `Dragonskin Belt of Brilliance`,
-    `Dragonskin Belt of Detection`,
-    `Dragonskin Belt of Giants`,
-    `Dragonskin Belt of Perfection`,
-    `Hard Leather Belt of Rage`,
-    `Hard Leather Belt of Vitriol`,
-    `Leather Belt of Protection`,
-    `Mesh Belt of Giants`,
-    `Mesh Belt of Perfection`,
-    `Mesh Belt of Rage`,
-    `Ornate Belt of Titans`,
-    `Wool Sash of Anger`,
-    `Wool Sash of Enlightenment`,
-    `Dragonskin Belt of Fury`,
-    `Dragonskin Belt of Skill`,
-    `Dragonskin Belt of the Twins`,
-    `Hard Leather Belt of Power`,
-    `Hard Leather Belt of Skill`,
-    `Leather Belt of Brilliance`,
-    `Leather Belt of Detection`,
-    `Leather Belt of the Fox`,
-    `Linen Sash of Perfection`,
-    `Mesh Belt of Anger`,
-    `Mesh Belt of Detection`,
-    `Mesh Belt of Reflection`,
-    `Mesh Belt of Vitriol`,
-    `Ornate Belt of Rage`,
-    `Ornate Belt of the Twins`,
-    `Plated Belt of Brilliance`,
-    `Silk Sash of Brilliance`,
-    `Silk Sash of Protection`,
-    `Silk Sash of Titans`,
-    `Wool Sash of Brilliance`,
-    `Brightsilk Sash of Vitriol`,
-    `Brightsilk Sash of the Fox`,
-    `Demonhide Belt of Fury`,
-    `Dragonskin Belt of Power`,
-    `Leather Belt of Skill`,
-    `Leather Belt of Titans`,
-    `Linen Sash of Detection`,
-    `Linen Sash of Skill`,
-    `Mesh Belt of Enlightenment`,
-    `Mesh Belt of the Fox`,
-    `Ornate Belt of Giants`,
-    `Ornate Belt of Power`,
-    `Plated Belt of Enlightenment`,
-    `Sash of Rage`,
-    `War Belt of Detection`,
-    `Dragonskin Belt of Titans`,
-    `Hard Leather Belt of Perfection`,
-    `Hard Leather Belt of Protection`,
-    `Hard Leather Belt of the Twins`,
-    `Linen Sash of the Twins`,
-    `Mesh Belt of Fury`,
-    `Ornate Belt of Anger`,
-    `Ornate Belt of Enlightenment`,
-    `Plated Belt of Fury`,
-    `Sash of Protection`,
-    `Silk Sash of Fury`,
-    `War Belt of Skill`,
-  ],
-
-  weapon: [
-    `Grave Wand`,
-    `Quarterstaff`,
-    `Falchion`,
-    `Tome`,
-    `Katana`,
-    `Scimitar`,
-    `Maul`,
-    `Short Sword`,
-    `Grimoire`,
-    `Chronicle`,
-    `Book`,
-    `Long Sword`,
-    `Wand`,
-    `Mace`,
-    `Bone Wand`,
-    `Ghost Wand`,
-    `Warhammer`,
-    `Club`,
-    `Falchion of Fury`,
-    `Falchion of Reflection`,
-    `Mace of Vitriol`,
-    `Bone Wand of Power`,
-    `Warhammer of Titans`,
-    `Chronicle of Detection`,
-    `Mace of Giants`,
-    `Mace of the Twins`,
-    `Chronicle of Rage`,
-    `Falchion of the Fox`,
-    `Mace of Skill`,
-    `Warhammer of Power`,
-    `Bone Wand of Fury`,
-    `Bone Wand of Titans`,
-    `Chronicle of Protection`,
-    `Mace of Detection`,
-    `Chronicle of the Twins`,
-    `Short Sword of the Twins`,
-    `Warhammer of Anger`,
-    `Bone Wand of Reflection`,
-    `Book of Protection`,
-    `Falchion of Enlightenment`,
-    `Mace of Brilliance`,
-    `Bone Wand of the Fox`,
-    `Chronicle of Giants`,
-    `Chronicle of Vitriol`,
-    `Falchion of Power`,
-    `Grimoire of Reflection`,
-    `Mace of Protection`,
-    `Short Sword of Skill`,
-    `Tome of the Fox`,
-    `Warhammer of Fury`,
-    `Warhammer of Perfection`,
-    `Warhammer of Reflection`,
-    `Katana of Brilliance`,
-    `Short Sword of Protection`,
-    `Warhammer of the Fox`,
-    `Bone Wand of Enlightenment`,
-    `Bone Wand of Perfection`,
-    `Club of Reflection`,
-    `Falchion of Anger`,
-    `Falchion of Titans`,
-    `Long Sword of Power`,
-    `Long Sword of the Fox`,
-    `Mace of Rage`,
-    `Short Sword of Brilliance`,
-    `Short Sword of Giants`,
-    `Short Sword of Rage`,
-    `Book of Giants`,
-    `Grave Wand of Protection`,
-    `Grimoire of Power`,
-    `Katana of Skill`,
-    `Short Sword of Detection`,
-    `Short Sword of Vitriol`,
-    `Wand of Brilliance`,
-    `Wand of Rage`,
-    `Bone Wand of Anger`,
-    `Club of Perfection`,
-    `Ghost Wand of Titans`,
-    `Grimoire of Enlightenment`,
-    `Grimoire of the Fox`,
-    `Long Sword of Reflection`,
-    `Quarterstaff of Protection`,
-    `Scimitar of Giants`,
-    `Tome of Power`,
-    `Wand of Skill`,
-    `Book of Rage`,
-    `Chronicle of Brilliance`,
-    `Club of Anger`,
-    `Club of Power`,
-    `Ghost Wand of Anger`,
-    `Grave Wand of Giants`,
-    `Katana of Protection`,
-    `Long Sword of Fury`,
-    `Maul of Fury`,
-    `Maul of Reflection`,
-    `Quarterstaff of Brilliance`,
-    `Quarterstaff of Vitriol`,
-    `Tome of Enlightenment`,
-    `Tome of Fury`,
-    `Wand of Protection`,
-    `Warhammer of Enlightenment`,
-  ],
-}
-
-export default loot
