--- src/constants/metadata/mutagen.ts
@@ -1,207 +1,0 @@
-const mutagen = {
-  Artist: ["Ilja Karilampi", "Tommy Cash", "Nik Kosmas", "Katja Novitskova"],
-  Generation: [
-    "Generation #0",
-    "Generation #1",
-    "Generation #2",
-    "Generation #3",
-    "Generation #7",
-    "Generation #14",
-    "Generation #4",
-    "Generation #11",
-    "Generation #23",
-    "Generation #5",
-    "Generation #10",
-    "Generation #12",
-    "Generation #43",
-    "Generation #6",
-    "Generation #8",
-  ],
-  Genesis: [
-    "Genesis #0",
-    "Genesis #29",
-    "Genesis #10",
-    "Genesis #28",
-    "Genesis #18",
-    "Genesis #38",
-    "Genesis #23",
-    "Genesis #6",
-    "Genesis #22",
-    "Genesis #7",
-    "Genesis #2",
-    "Genesis #14",
-    "Genesis #3",
-    "Genesis #35",
-    "Genesis #1",
-    "Genesis #16",
-    "Genesis #27",
-    "Genesis #32",
-    "Genesis #34",
-    "Genesis #8",
-    "Genesis #12",
-    "Genesis #17",
-    "Genesis #19",
-    "Genesis #20",
-    "Genesis #21",
-    "Genesis #26",
-    "Genesis #33",
-    "Genesis #13",
-    "Genesis #24",
-    "Genesis #25",
-    "Genesis #31",
-    "Genesis #36",
-    "Genesis #9",
-    "Genesis #11",
-    "Genesis #15",
-    "Genesis #30",
-    "Genesis #37",
-    "Genesis #39",
-    "Genesis #4",
-    "Genesis #5",
-  ],
-  Layer: [0, 1, 2, 3],
-  "Mutation type": ["Regular", "Rare"],
-  Rarity: ["Regular", "Golden"],
-  Type: ["Mutagen", "Print", "Genesis"],
-  "Variant number": [
-    74, 37, 87, 62, 94, 7, 52, 14, 76, 82, 0, 48, 3, 48, 68, 80, 88, 9, 48, 90, 25,
-    54, 89, 93, 95, 17, 21, 4, 46, 79, 22, 38, 49, 77, 13, 23, 53, 63, 64, 96, 99,
-    33, 83, 10, 27, 41, 42, 43, 72, 78, 29, 36, 52, 57, 59, 61, 67, 86, 97, 16, 19,
-    44, 20, 81, 92, 12, 15, 18, 51, 58, 6, 38, 98, 2, 37, 28, 31, 46, 48, 55, 75, 91,
-    26, 39, 40, 60, 70, 71, 32, 34, 65, 73, 85, 30, 56, 66, 24, 45, 47, 69, 84, 1,
-    30, 11, 5, 30, 50, 8, 30, 35,
-  ],
-  Property: [
-    50,
-    "2CHAINZ",
-    "BUY-LOW-SELL-HIGH",
-    "TRANSPARENT",
-    "00FFFF-0000FF",
-    "0000FF-00FF00",
-    "MOON",
-    "FFFF00-00FFFF",
-    "FF00FF-FFFF00",
-    "MUTAGEN",
-    "FF0000-FF00FF",
-    "SHROOMS",
-    "00FF00-FF0000",
-    "CATS",
-    "ICECUBE",
-    "MINT",
-    "SHOJO",
-    "00-MUTAGEN",
-    "01-MUTAGEN",
-    "02-MUTAGEN",
-    "03-MUTAGEN",
-    "DNA-03-04",
-    "04-MUTAGEN",
-    "05-MUTAGEN",
-    "CENTAURESS",
-    "DNA-00-01",
-    "DNA-00-05",
-    "DNA-01-02",
-    "DNA-02-03",
-    "DNA-04-05",
-    "ETHEREUM",
-    "FAIRY",
-    "SOLIDITY",
-    "UNICORN",
-    "PEARL",
-    "2021",
-    "25",
-    "ANGEL",
-    "APE",
-    "BALL",
-    "BEAR-MARCH",
-    "BIN",
-    "BLOOD",
-    "BLUE-CRAB",
-    "BUBBLE-GUM",
-    "BULL",
-    "BUTTERFLY",
-    "CASTLE",
-    "CENTAUR",
-    "CHAN",
-    "CLOUD",
-    "CRYBABY",
-    "DEMON GIRL",
-    "DIVER",
-    "DNA-00-02",
-    "DNA-00-03",
-    "DNA-01-04",
-    "DOG",
-    "DRAGON",
-    "DROP",
-    "DUOBRACHIUM",
-    "EAGLE-2",
-    "EAGLE-3",
-    "EAGLE-LITE",
-    "EMPEROR-DUMBO",
-    "EN",
-    "ETH",
-    "EURYPHARYNX",
-    "FACES",
-    "FALCON-1",
-    "FALCON-9",
-    "FALCON-9v11",
-    "FALCON-HEAVY",
-    "FINGER-CHAMELEON",
-    "FIRE",
-    "FLAMING-HEART",
-    "FLY",
-    "FRUIT-BASKET",
-    "G",
-    "64",
-    "GHOST",
-    "GHOST-PLANT",
-    "HOLE",
-    "HORNERA-MEDITERRANEA",
-    "HORSE",
-    "ILJA-KARILAMPI",
-    "INDIFFERENCE",
-    "JELLYFISH",
-    "JO",
-    "KATJA-NOVITSKOVA",
-    "MERMAID-HAIR",
-    "MU",
-    "NIK-KOSMAS",
-    "NIKA",
-    "NINJA-LANTERNSHARK",
-    "NYAA",
-    "OCTOPUS",
-    "OK",
-    "ORBIT",
-    "PELICAN",
-    "PEPE",
-    "ROAD-TO-RIVER",
-    "ROSE",
-    "RUN",
-    "SAFE-FART",
-    "SALUTE",
-    "SHELL",
-    "SKY",
-    "SLIME",
-    "SOLAR",
-    "SPHINX",
-    "SPLASH",
-    "STARS",
-    "SUPERFALCON",
-    "SUSHI",
-    "TA",
-    "TARDIGRADE",
-    "TIDE",
-    "TOMMY-CASH",
-    "TREE-OF-LIFE",
-    "UNDERWATER-FART",
-    "VERAX",
-    "VITALIK",
-    "WATERFALL",
-    "WOLF",
-    "YAWN",
-    "YETI-CRAB",
-    "Moon",
-    "Punk",
-  ],
-}
-
-export default mutagen
