+++ src/pages/[guild]/add-role.tsx
@@ -5,11 +5,11 @@ import ConnectWalletAlert from "components/common/ConnectWalletAlert"
 import ErrorAnimation from "components/common/ErrorAnimation"
 import Layout from "components/common/Layout"
 import Section from "components/common/Section"
+import Description from "components/create-guild/Description"
 import LogicPicker from "components/create-guild/LogicPicker"
+import NameAndIcon from "components/create-guild/NameAndIcon"
 import Requirements from "components/create-guild/Requirements"
-import Description from "components/create/Description"
-import NameAndIcon from "components/create/NameAndIcon"
-import SubmitButton from "components/create/SubmitButton"
+import SubmitButton from "components/create-guild/SubmitButton"
 import useGuild from "components/[guild]/hooks/useGuild"
 import useWarnIfUnsavedChanges from "hooks/useWarnIfUnsavedChanges"
 import { useEffect, useState } from "react"
