+++ src/pages/[guild]/edit/index.tsx
@@ -16,10 +16,10 @@ import { useWeb3React } from "@web3-react/core"
 import CtaButton from "components/common/CtaButton"
 import Layout from "components/common/Layout"
 import Section from "components/common/Section"
+import Description from "components/create-guild/Description"
 import LogicPicker from "components/create-guild/LogicPicker"
+import NameAndIcon from "components/create-guild/NameAndIcon"
 import Requirements from "components/create-guild/Requirements"
-import Description from "components/create/Description"
-import NameAndIcon from "components/create/NameAndIcon"
 import DeleteGuildCard from "components/[guild]/edit/index/DeleteGuildCard"
 import useEdit from "components/[guild]/EditButtonGroup/components/CustomizationButton/hooks/useEdit"
 import useGuild from "components/[guild]/hooks/useGuild"
