+++ src/pages/create-guild.tsx
@@ -6,12 +6,12 @@ import ConnectWalletAlert from "components/common/ConnectWalletAlert"
 import ErrorAnimation from "components/common/ErrorAnimation"
 import Layout from "components/common/Layout"
 import Section from "components/common/Section"
+import Description from "components/create-guild/Description"
 import LogicPicker from "components/create-guild/LogicPicker"
+import NameAndIcon from "components/create-guild/NameAndIcon"
 import PickRolePlatform from "components/create-guild/PickRolePlatform"
 import Requirements from "components/create-guild/Requirements"
-import Description from "components/create/Description"
-import NameAndIcon from "components/create/NameAndIcon"
-import SubmitButton from "components/create/SubmitButton"
+import SubmitButton from "components/create-guild/SubmitButton"
 import useWarnIfUnsavedChanges from "hooks/useWarnIfUnsavedChanges"
 import { useEffect, useState } from "react"
 import { FormProvider, useForm, useWatch } from "react-hook-form"
