+++ src/static/avatars/addressAvatarPairs.ts
